HomeWork#3 fix 3 bugs
