+++ kanaievProj/src/test/java/loginTest/LoginTestWithPageObject.java
@@ -17,32 +17,13 @@ public class LoginTestWithPageObject extends BaseTest {
 
     @Test
     public void invalidLogin() {
-        loginPage.openLoginPage();
-        loginPage.enterLoginInSignIn("auto");
-        loginPage.enterPasswordInSignIn("123");
-        loginPage.clickOnButtonSignIn();
+        loginPage.fillLoginFormAndSubmit("auto", "123");
 
-        checkExpectedResult("Button SingOut is visible",
+        checkExpectedResult("Button SingOut is visible, but shouldn`t",
                 homePage.isButtonSignOutPresent(), false);
         checkExpectedResult("Button SingIn is not visible",
                 loginPage.isButtonSignInPresent(), true);
         checkExpectedResult("Alert Invalid Sign In login is not visible",
                 loginPage.isAlertInvalidSingInPresent(), true);
-    }
-
-    @Test
-    public void validationOnRegistration(){
-        loginPage.openLoginPage();
-        loginPage.enterLoginInRegLogin("tr");
-        loginPage.enterEmailInRegEmail("test.com");
-        loginPage.enterPasswordInRegPassword("123");
-        loginPage.clickOnButtonSignUp();
-
-        checkExpectedResult("Alert invalid Sing Up login is not visible",
-                loginPage.isAlertInvalidSingUpLoginPresent(), true);
-        checkExpectedResult("Alert invalid Sing Up email is not visible",
-                loginPage.isAlertInvalidSingUpEmailPresent(), true);
-        checkExpectedResult("Alert invalid Sing Up password is not visible",
-                loginPage.isAlertInvalidSingUpPasswordPresent(), true);
     }
 }
